API/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
mongoengine==0.24.1
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
pyjwt==2.6.0

